+++ templates/index.twig
@@ -336,20 +336,22 @@
 				{% set sectionId = '02' %}
 				{{ include('/sections/title-switcher.twig') }}
 				{% set sectionId = '03' %}
-				{{ include('/sections/featured-slider.twig') }}
+				{{ include('/sections/supply-chain-animation.twig') }}
 				{% set sectionId = '04' %}
-				{{ include('/sections/tabs.twig') }}
+				{{ include('/sections/featured-slider.twig') }}
 				{% set sectionId = '05' %}
-				{{ include('/sections/features.twig') }}
+				{{ include('/sections/tabs.twig') }}
 				{% set sectionId = '06' %}
-				{{ include('/sections/logos-slider.twig') }}
+				{{ include('/sections/features.twig') }}
 				{% set sectionId = '07' %}
-				{{ include('/sections/events-slider.twig') }}
+				{{ include('/sections/logos-slider.twig') }}
 				{% set sectionId = '08' %}
-				{{ include('/sections/latest-podcasts.twig') }}
+				{{ include('/sections/events-slider.twig') }}
 				{% set sectionId = '09' %}
-				{{ include('/sections/latest-news.twig') }}
+				{{ include('/sections/latest-podcasts.twig') }}
 				{% set sectionId = '10' %}
+				{{ include('/sections/latest-news.twig') }}
+				{% set sectionId = '11' %}
 				{{ include('/sections/outro.twig') }}
 				{% set sectionId = '' %}
 			{% endblock %}
