+++ templates/sections/features.twig
@@ -18,8 +18,8 @@
 			{% endif %}
             <div class="features__row">
                 <div class="features__col features__col--animation">
-                    <div class="features__animation features__animation--desktop js-homepage-animation" data-animation-path="{% if (craft.app.sites.primarySite.baseUrl == 'http://wiliot.adchitects.co/web/') %}/web{% endif %}/src/img/homepage-animation.json?v=07112022"></div>
-                    <div class="features__animation features__animation--mobile js-homepage-animation" data-animation-path="{% if (craft.app.sites.primarySite.baseUrl == 'http://wiliot.adchitects.co/web/') %}/web{% endif %}/src/img/homepage-animation-mobile.json?v=07112022"></div>
+                    <div class="features__animation features__animation--desktop js-homepage-animation" data-rwd="desktop" data-animation-path="{% if (craft.app.sites.primarySite.baseUrl == 'http://wiliot.adchitects.co/web/') %}/web{% endif %}/src/img/homepage-animation.json?v=07112022"></div>
+                    <div class="features__animation features__animation--mobile js-homepage-animation" data-rwd="mobile" data-animation-path="{% if (craft.app.sites.primarySite.baseUrl == 'http://wiliot.adchitects.co/web/') %}/web{% endif %}/src/img/homepage-animation-mobile.json?v=07112022"></div>
                 </div>
                 {% for content in contentList %}
 					{% set title = content.featureContentTitle %}
